Support MEMBER OF on entities with association IDs
Currently, SqlWalker assumes that entities in a MEMBER OF have a regular field as join column. When an association is used as join, an Undefined index is thrown because $class->fieldNames['association join column name'] does not exist. This commit fixes this by looking up the column name of the relevant join column instead.