Adapt test logic to PHP and SQLite II
In a88242ee6ceb1dc8477b715c20b2f2f5e4fdfa28, testDateSub() was modified
in order to satisfy different opinions to the question "What is now
minus one month", that has different answers for different pieces of
software on March 30th.

Today, we are January 30th, we need to do the same for testDateAdd().
Hopefully this is the last time we hear about this.