Fix exception message in "missingFieldName" MappingException method
For XML Mapping, the exception message in case an association field isn't configured properly states that it misses the "fieldName" attribute.
But even if "fieldName" is indeed a key in the mapping associative array in ClassMetadataInfo, it is read from the "field" XML attribute in XmlDriver.

This commit changes the exception message to suggest the right correction to apply when this exception occurs.