Only put `collation` on join columns, never `charset`
The fix introduced in #9636 now breaks when working with integer relations.

Putting `charset` and `collation` on a text join column works fine, but when the
column is an integer, MySQL won't accept it:

```sql
CREATE TABLE table_name (
  // ..
  user_id INT CHARACTER SET utf8 NOT NULL COLLATE `utf8_unicode_ci`
  // ..
)
```

produces this error:
```
Query 1 ERROR: You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near 'CHARACTER SET utf8 NOT NULL COLLATE `utf8_unicode_ci`, type VARCHAR(255) NOT NUL' at line 1
```

Luckily the solution is simple. The charset is not needed as MySQL automatically derives that from
the `collation` when specified. It's also clever enough to ignore `collation` when the type is an integer.

Therefore the following works fine:
```sql
CREATE TABLE table_name (
  // ..
  user_id INT NOT NULL COLLATE `utf8_unicode_ci`,
  string_user_id VARCHAR(255) NOT NULL COLLATE `utf8_unicode_ci`,
  // ..
)
```