Added parameter to getSelectJoinColumnSQL
The Collection of a One-To-Many Association is empty when using a
junction entity having Class Table Inheritance to connect 2 entities.
The problem arises because the Collection will only be hydrated if the
association mappings have an identifier, but the identifier is
not detected because $isIdentifierColumn is hardcoded to false.
Fixes #9822