Fix persisting with class table inheritance
This PR fixes the JoinedSubclassPersister throwing a LengthException.
This happens as it tries to fetch the entitie's generated columns
as well as a potentially existing version column before the entity has
been persisted.

The fix is quite simple: It just moves the call to the
assignDefaultVersionAndUpsertableValues() method so it is executed
after the the actual change/insert.

Closes #9378
Fixes #9378