Failing test: Colliding column definitions in STI
When using Single Table Inheritance and two sibling classes declare fields with the same name, but different column definitions (say, one is `int`, the other one `string`), the resulting conflict is not raised as a problem.

 #### Current situation

The database schema will be created based on one of the two definitions. In the current example, we actually end up with an `INT` column. Values can not be persisted correctly and are lost.

 #### Expected behaviour

The ORM should either refuse this configuration, clearly pointing out what the problem is, an possibly suggest to use `name="..."` declarations to make sure different columns for the different fields can co-exist in the database.

Or the naming strategy (or whatever mechanism) can use pre/suffixes automatically to resolve the situation, albeit that would probably be a larger BC break and would cause old schemas to not work anymore.

We might consider to check if the column definitions actually differ and use a single column for two fields in two classes if the definitions match. In the long run, that might be hard to maintain – for example, what should happen once one of the two definitions changes?

 #### Additional information

Seems to work well for JTI, since different fields end up in different tables and each one gets the right column definition.